+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/model/upsert/output/OutputUpsertData.kt
@@ -22,6 +22,8 @@ package org.eclipse.tractusx.bpdm.gate.model.upsert.output
 import org.eclipse.tractusx.bpdm.common.dto.AddressType
 import org.eclipse.tractusx.bpdm.common.dto.BusinessPartnerRole
 import org.eclipse.tractusx.bpdm.gate.api.model.BusinessPartnerScriptVariantDto
+import org.eclipse.tractusx.orchestrator.api.model.AddressGoldenRecordRelation
+import org.eclipse.tractusx.orchestrator.api.model.LegalEntityGoldenRecordRelation
 
 data class OutputUpsertData(
     val nameParts: List<String>,
@@ -43,5 +45,7 @@ data class OutputUpsertData(
     val legalEntityConfidence: ConfidenceCriteria,
     val siteConfidence: ConfidenceCriteria?,
     val addressConfidence: ConfidenceCriteria,
-    val scriptVariants: List<BusinessPartnerScriptVariantDto>
+    val scriptVariants: List<BusinessPartnerScriptVariantDto>,
+    val legalEntityGoldenRecordRelations: List<LegalEntityGoldenRecordRelation>,
+    val addressGoldenRecordRelations: List<AddressGoldenRecordRelation>
 )